fastapi_ai/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
numpy==1.26.4
pydantic==2.6.3

# ML — ikisinden birini seç, diğerini yorum satırı yap:
# tensorflow==2.15.0
# torch==2.2.0
# torchvision==0.17.0
